Keep saga state internal
